[김희진] sprint7 (#96)
* refactor: code review

* feat: add item detail router and page

* feat: add PrimaryRoundButton, check onDelete in TagCard

* feat: add Favorite, ItemDetails components

* feat: ItemDetails

* fix: item createdAt wrapper css

* feat: add item detail enquiry

* feat: add itemDetail responsive css

* feat: add enquiry button onChange

* feat: add item Comments

* feat: add comments

* style: remove console log

* refactor: remove useless comparing

* refactor

+++ src/components/item/TagCard.jsx
@@ -3,13 +3,21 @@ import deleteIcon from '../../assets/ic_X.svg';
 
 function TagCard({ index, name, onDelete }) {
   const handleOnClick = () => {
-    onDelete(index);
+    if (onDelete) onDelete(index);
   };
 
   return (
     <div className="TagCard">
       <span>#{name}</span>
-      <img src={deleteIcon} alt="삭제 아이콘" onClick={handleOnClick} />
+      {onDelete && (
+        <button
+          className="TagCard-btn"
+          onClick={handleOnClick}
+          aria-label="삭제"
+        >
+          <img src={deleteIcon} alt="삭제 아이콘" />
+        </button>
+      )}
     </div>
   );
 }
